requirements.txt
watchdog
paramiko
python-dotenv
PyYAML
